+++ awswl/__init__.py
@@ -1,3 +1,3 @@
 from usingversion import getattr_with_version
 
-__getattr__ = getattr_with_version("mypackage", __file__, __name__)
+__getattr__ = getattr_with_version("awswl", __file__, __name__)